+++ CHANGES.rst
@@ -4,7 +4,8 @@ Changelog
 1.3.15 (unreleased)
 -------------------
 
-- Nothing changed yet.
+- Fix double quoted email address validation.
+  [chris-adam]
 
 
 1.3.14 (2026-04-14)